created elements.js
moved css to elements.js
