+++ src/components/forms/FormLimit.js
@@ -1,57 +1,11 @@
 import React, { useState } from "react";
-import styled from "styled-components";
-
-const InputContainer = styled.div`
-    display: grid;
-  align-items: center;
-  justify-items: center;
-  justify-content: center;
-  padding: 0.1rem 0.3rem;
-  background-color: hsla(40, 90%, 50%, 1);
-  width: 95%;
-`;
-
-const InputBox = styled.form`
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(4, auto);
-  grid-auto-flow: row;
-  align-items: center;
-  align-content: space-around;
-  justify-content: center;
-  padding: 0.1rem 0.3rem;
-  background-color: hsla(40, 90%, 50%, 1);
-`;
-
-const InputField = styled.input`
- width: 40%;
-  margin: 0;
-  padding: 0.1rem 0.3rem;
-  font-size: 1rem;
-  font-weight: 600;
-  color: hsla(70, 30%, 30%, 1);
-  background-color: #fff;
-`;
-
-const InputLabel= styled.label`
-height: 100%;
-  margin: 0;
-  padding: 0.1rem 0.3rem;
-  font-size: 1rem;
-  font-weight: 600;
-  text-align: right;
-  color: #fff;
-`;
-
-const SubmitButton= styled.button`
-  grid-column-start: 4;
-  grid-column-end: 5;
-  padding: 0.3rem 1rem;
-  font-size: 1rem;
-  font-weight: 900;
-  background-color: var(--green);
-  color: #fff;
-`;
+import {
+  InputContainer,
+  InputBox,
+  InputField,
+  InputLabel,
+  SubmitButton
+} from "../../styles/elements.js";
 
 const FormLimit = props => {
   const [itemPrice, setPrice] = useState(200);
@@ -121,9 +75,7 @@ const FormLimit = props => {
           value={itemQuantity}
           onChange={e => setQuantity(e.target.value)}
         />
-        <SubmitButton type="submit">
-        Uložit
-        </SubmitButton>
+        <SubmitButton type="submit">Uložit</SubmitButton>
       </InputBox>
     </InputContainer>
   );
