+++ src/components/panels/BarStripe.js
