+++ src/components/panels/SwitcherPanel.js
@@ -1,5 +1,6 @@
 import React from "react";
-import CalculatePanel from "../components/CalculatePanel";
+import CalculatePanel from "./CalculatePanel";
+import {BasicButton} from '../../styles/elements.js'
 import styled from "styled-components";
 
 const SwitchPanel = styled.div`
@@ -9,32 +10,6 @@ const SwitchPanel = styled.div`
   justify-items: space-around;
   padding: 0.5rem 0.3rem;
   background: hsla(24, 90%, 60%, 1);
-`;
-
-const BasicButton = styled.button`
-  padding: 0.3rem 1rem;
-  font-size: 2rem;
-  font-weight: 900;
-  background-color: var(--green);
-  color: #fff;
-  z-index: 4;
-  transition: all 200ms cubic-bezier(0.215, 0.61, 0.355, 1);
-  &:hover {
-    color: var(--green);
-    background-color: white;
-  }
-  @media all and (max-width: 1680px) {
-    font-size: 1.5rem;
-  }
-  @media all and (max-width: 980px) {
-    font-size: 1.3rem;
-  }
-  @media all and (max-width: 736px) {
-    font-size: 1.2rem;
-  }
-  @media all and (max-width: 480px) {
-    font-size: 1.1rem;
-  }
 `;
 
 export default function SwitcherPanel(props) {
