+++ src/pages/Login.js
@@ -1,27 +1,14 @@
 import React from "react";
-import AuthForm from "../auth/AuthForm";
+import AuthForm from "../components/forms/AuthForm";
 import Navigation from "../components/Navigation";
-import styled from 'styled-components'
-
-const PageLayout = styled.div`
-display: grid;
-grid-auto-flow: row;
-justify-items: center;
-margin: 0;
-padding: 0;
-min-height: 100vh;
-overflow: hidden;
-@media all and (max-width: 480px) {
-    font-size: 1.1rem;
-  }
-`;
+import { PageLayout } from "../styles/elements.js";
 
 const Login = () => {
   return (
     <PageLayout>
       <Navigation />
-      <AuthForm register={false}/>     
-      </PageLayout>
+      <AuthForm register={false} />
+    </PageLayout>
   );
 };
